refactor(web): searchInput moved to a folder
