+++ packages/web/src/components/components.ts
@@ -25,5 +25,6 @@ export * from './LoadingOverlay'
 export * from './InputBase'
 export * from './Switch'
 export * from './NumberIncrement'
+export * from './SearchInput'
 
 export * from './defaultStyles'
